Split method identifier
