+++ src/CMakeLists.txt
@@ -153,6 +153,8 @@ libcvc5_add_sources(
   proof/lazy_proof_chain.h
   proof/lazy_tree_proof_generator.cpp
   proof/lazy_tree_proof_generator.h
+  proof/method_id.cpp
+  proof/method_id.h
   proof/proof.cpp
   proof/proof.h
   proof/proof_checker.cpp
